backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pyjwt==2.10.1
slowapi==0.1.9
httpx==0.28.1
minio==7.2.12
python-multipart==0.0.17
pydantic-settings==2.6.1
